refactor: resructure folders
