+++ src/image_embedding/mod.rs
@@ -7,5 +7,4 @@ mod utils;
 mod init;
 pub use init::*;
 
-mod models;
-pub use models::*;
+mod r#impl;
