+++ src/sparse_text_embedding/init.rs
@@ -1,0 +1,47 @@
+use std::path::{Path, PathBuf};
+
+use ort::{ExecutionProviderDispatch, Session};
+use tokenizers::Tokenizer;
+
+use crate::{models::sparse::SparseModel, TokenizerFiles, DEFAULT_CACHE_DIR};
+
+use super::{DEFAULT_EMBEDDING_MODEL, DEFAULT_MAX_LENGTH};
+
+/// Options for initializing the SparseTextEmbedding model
+#[derive(Debug, Clone)]
+pub struct SparseInitOptions {
+    pub model_name: SparseModel,
+    pub execution_providers: Vec<ExecutionProviderDispatch>,
+    pub max_length: usize,
+    pub cache_dir: PathBuf,
+    pub show_download_progress: bool,
+}
+
+impl Default for SparseInitOptions {
+    fn default() -> Self {
+        Self {
+            model_name: DEFAULT_EMBEDDING_MODEL,
+            execution_providers: Default::default(),
+            max_length: DEFAULT_MAX_LENGTH,
+            cache_dir: Path::new(DEFAULT_CACHE_DIR).to_path_buf(),
+            show_download_progress: true,
+        }
+    }
+}
+
+/// Struct for "bring your own" embedding models
+///
+/// The onnx_file and tokenizer_files are expecting the files' bytes
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct UserDefinedSparseModel {
+    pub onnx_file: Vec<u8>,
+    pub tokenizer_files: TokenizerFiles,
+}
+
+/// Rust representation of the SparseTextEmbedding model
+pub struct SparseTextEmbedding {
+    pub tokenizer: Tokenizer,
+    pub(crate) session: Session,
+    pub(crate) need_token_type_ids: bool,
+    pub(crate) model: SparseModel,
+}
