+++ tests/assets/image_0.png
